+++ packages/sources/cfbenchmarks/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/cfbenchmarks-adapter
 
+## 2.6.0
+
+### Minor Changes
+
+- [#4231](https://github.com/smartcontractkit/external-adapters-js/pull/4231) [`8d06b4a`](https://github.com/smartcontractkit/external-adapters-js/commit/8d06b4ab8f64e0c5ecbf8e69e5a1786dc0061d77) Thanks [@mxiao-cll](https://github.com/mxiao-cll)! - Bump market status dependency
+
+- [#4213](https://github.com/smartcontractkit/external-adapters-js/pull/4213) [`a520a2d`](https://github.com/smartcontractkit/external-adapters-js/commit/a520a2d0fae0680280cdf3f1bc9182b9b679b965) Thanks [@mmcallister-cll](https://github.com/mmcallister-cll)! - Fix lwba cache fetch when api secondary enabled
+
+### Patch Changes
+
+- [#4198](https://github.com/smartcontractkit/external-adapters-js/pull/4198) [`bf10006`](https://github.com/smartcontractkit/external-adapters-js/commit/bf100063b8068e1ce0f3a3e4d410a0babc61c2f9) Thanks [@mmcallister-cll](https://github.com/mmcallister-cll)! - framework dependency bump
+
 ## 2.5.2
 
 ### Patch Changes